Factoring out temperature behavior for MOS
